// src/constants/enums.ts
export enum FeatureCategory {
    Performance = "Performance",
    Comfort = "Comfort",
    Design = "Design",
    Safety = "Safety",
    Technology = "Technology",
    Convenience = "Convenience",
    HeavyDuty="Heavy Duty"
}

export enum VehicleColors {
    PristineWhite = "Pristine White",
    NeptuneBlue = "Neptune Blue"
}

export enum ModeType {
    AUTO = "Auto",
    MANUAL = "Manual"
}